Fix build.

+++ src/libs/vgeometry/vlayoutplacelabel.cpp
@@ -35,7 +35,7 @@
 // See https://stackoverflow.com/a/46719572/3045403
 #if __cplusplus < 201703L // C++17
 constexpr quint32 VLayoutPlaceLabel::streamHeader;  // NOLINT(readability-redundant-declaration)
-constexpr quint32 VLayoutPlaceLabel::classVersion;  // NOLINT(readability-redundant-declaration)
+constexpr quint16 VLayoutPlaceLabel::classVersion;  // NOLINT(readability-redundant-declaration)
 #endif
 
 //---------------------------------------------------------------------------------------------------------------------
