+++ src/libs/vlayout/vlayoutpiece_p.h
@@ -132,7 +132,7 @@ QT_WARNING_POP
 // See https://stackoverflow.com/a/46719572/3045403
 #if __cplusplus < 201703L // C++17
 constexpr quint32 VLayoutPieceData::streamHeader;  // NOLINT(readability-redundant-declaration)
-constexpr quint32 VLayoutPieceData::classVersion;  // NOLINT(readability-redundant-declaration)
+constexpr quint16 VLayoutPieceData::classVersion;  // NOLINT(readability-redundant-declaration)
 #endif
 
 // Friend functions
